requirements.txt
vnstock
requests
pandas
flask
Flask-APScheduler
flask-sqlalchemy 
pymysql
python-dotenv
uWSGI
Flask-Caching
redis
# celery
pika